@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp1.1</TargetFramework>
+    <TargetFramework>netcoreapp2.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
@@ -12,6 +12,7 @@
     <PackageReference Include="Microsoft.ApplicationInsights.AspNetCore" Version="2.0.0" />
     <PackageReference Include="Microsoft.AspNetCore" Version="1.1.1" />
     <PackageReference Include="Microsoft.AspNetCore.Mvc" Version="1.1.2" />
+    <PackageReference Include="Microsoft.NETCore.Portable.Compatibility" Version="1.0.1" />
     <PackageReference Include="NLog.Web.AspNetCore" Version="4.3.1" />
   </ItemGroup>
 
@@ -20,14 +21,20 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\ClassLibrary1\ClassLibrary1.csproj" />
+  </ItemGroup>
 
+  <ItemGroup>
     <Reference Include="AlexaSkillsKit">
-      <HintPath>..\..\..\..\..\Desktop\AlexaSkillsKit.dll</HintPath>
+      <HintPath>..\..\..\..\Desktop\AlexaSkillsKit.NET\AlexaSkillsKit.Lib\bin\Debug\AlexaSkillsKit.dll</HintPath>
     </Reference>
-
     <Reference Include="BouncyCastle.Crypto">
-      <HintPath>..\..\..\..\..\Desktop\AlexaSkillsKit.NET\AlexaSkillsKit.Lib\bin\Debug\BouncyCastle.Crypto.dll</HintPath>
+      <HintPath>..\..\..\..\Desktop\AlexaSkillsKit.NET\AlexaSkillsKit.Lib\bin\Debug\BouncyCastle.Crypto.dll</HintPath>
+    </Reference>
+    <Reference Include="System.Net.Http.Formatting">
+      <HintPath>..\..\..\..\Desktop\AlexaSkillsKit.NET\AlexaSkillsKit.Lib\bin\Debug\System.Net.Http.Formatting.dll</HintPath>
     </Reference>
   </ItemGroup>
+
 
 </Project>
