@@ -8,177 +8,112 @@ using AlexaSkillsKit.UI;
 using AlexaSkillsKit.Slu;
 using Microsoft.Extensions.Logging;
 using NLog;
+using Newtonsoft.Json;
+using Microsoft.AspNetCore.Http;
+using System.IO;
+using BusInfo;
+
 
 namespace TransitKingSkill
 {
-    public class SampleSessionSpeechlet : Speechlet
+    public class SampleSessionSpeechlet : MySpeechletAsync
     {
-        private static Logger Log = LogManager.GetCurrentClassLogger();
+        private const string BusKey = "Bus";
+        private const string BusForIntent = "BusForIntent";
+        MyStopInfo _busInfo = new MyStopInfo(new BusLocator(), new TimeZoneConverter());
+        (string lat, string lon) conventionCenter = ("47.611959", "-122.332893");
 
-        // Note: NAME_KEY being a JSON property key gets camelCased during serialization
-        private const string NAME_KEY = "name";
-        private const string NAME_SLOT = "Name";
-
-
-        public override void OnSessionStarted(SessionStartedRequest request, Session session)
+        public override async Task<SpeechletResponse> OnIntentAsync(IntentRequest intentRequest, Session session)
         {
-            Log.Info("OnSessionStarted requestId={0}, sessionId={1}", request.RequestId, session.SessionId);
-        }
-
-
-        public override SpeechletResponse OnLaunch(LaunchRequest request, Session session)
-        {
-            Log.Info("OnLaunch requestId={0}, sessionId={1}", request.RequestId, session.SessionId);
-            return GetWelcomeResponse();
-        }
-
-
-        public override SpeechletResponse OnIntent(IntentRequest request, Session session)
-        {
-            Log.Info("OnIntent requestId={0}, sessionId={1}", request.RequestId, session.SessionId);
-
             // Get intent from the request object.
-            Intent intent = request.Intent;
-            string intentName = (intent != null) ? intent.Name : null;
+            var intent = intentRequest.Intent;
+            var intentName = intent?.Name;
 
             // Note: If the session is started with an intent, no welcome message will be rendered;
             // rather, the intent specific response will be returned.
-            if ("MyNameIsIntent".Equals(intentName))
+            if (BusForIntent.Equals(intentName))
             {
-                return SetNameInSessionAndSayHello(intent, session);
+                return await GetBusResponse(intent, session);
             }
-            else if ("WhatsMyNameIntent".Equals(intentName))
+
+            throw new SpeechletException("Invalid Intent");
+        }
+
+        private async Task<SpeechletResponse> GetBusResponse(Intent intent, Session session)
+        {
+            // Retrieve date from the intent slot
+            var busSlot = intent.Slots[BusKey];
+
+            string output;
+            if (busSlot != null)
             {
-                return GetNameFromSessionAndSayHello(intent, session);
+                string busRoute = busSlot.ToString();
+                var times = await _busInfo.GetArrivalTimesForRouteName(busRoute, conventionCenter.lat, conventionCenter.lon);
+
+                if (times != null)
+                {
+                    output = $"The next {busSlot} bus arrives at ";
+                    foreach (var time in times)
+                    {
+                        output += time;
+                    }
+                }
+                else
+                {
+                    output = $"Sorry, no bus info is available for {busSlot}";
+                }
             }
             else
             {
-                throw new SpeechletException("Invalid Intent");
+                output = $"Sorry, no bus info is available for {busSlot}";
             }
+
+            // Here we are setting shouldEndSession to false to not end the session and
+            // prompt the user for input
+            return BuildSpeechletResponse(intent.Name, output, false);
         }
 
-
-        public override void OnSessionEnded(SessionEndedRequest request, Session session)
+        public override Task<SpeechletResponse> OnLaunchAsync(LaunchRequest launchRequest, Session session)
         {
-            Log.Info("OnSessionEnded requestId={0}, sessionId={1}", request.RequestId, session.SessionId);
+            return Task.FromResult(GetWelcomeResponse());
         }
 
-
-        /**
-         * Creates and returns a {@code SpeechletResponse} with a welcome message.
-         * 
-         * @return SpeechletResponse spoken and visual welcome message
-         */
         private SpeechletResponse GetWelcomeResponse()
         {
-            // Create the welcome message.
-            string speechOutput =
-                "Welcome to the Alexa AppKit session sample app, please tell me your name by saying, my name is Sam";
-
-            // Here we are setting shouldEndSession to false to not end the session and
-            // prompt the user for input
-            return BuildSpeechletResponse("Welcome", speechOutput, false);
+            var output = "Welcome to the Transit King app. Please request a bus number near you.";
+            return BuildSpeechletResponse("Welcome", output, false);
         }
 
-
-        /**
-         * Creates a {@code SpeechletResponse} for the intent and stores the extracted name in the
-         * Session.
-         * 
-         * @param intent
-         *            intent for the request
-         * @return SpeechletResponse spoken and visual response the given intent
-         */
-        private SpeechletResponse SetNameInSessionAndSayHello(Intent intent, Session session)
-        {
-            // Get the slots from the intent.
-            Dictionary<string, Slot> slots = intent.Slots;
-
-            // Get the name slot from the list slots.
-            Slot nameSlot = slots[NAME_SLOT];
-            string speechOutput = "";
-
-            // Check for name and create output to user.
-            if (nameSlot != null)
-            {
-                // Store the user's name in the Session and create response.
-                string name = nameSlot.Value;
-                session.Attributes[NAME_KEY] = name;
-                speechOutput = String.Format(
-                    "Hello {0}, now I can remember your name, you can ask me your name by saying, whats my name?", name);
-            }
-            else
-            {
-                // Render an error since we don't know what the users name is.
-                speechOutput = "I'm not sure what your name is, please try again";
-            }
-
-            // Here we are setting shouldEndSession to false to not end the session and
-            // prompt the user for input
-            return BuildSpeechletResponse(intent.Name, speechOutput, false);
-        }
-
-
-        /**
-         * Creates a {@code SpeechletResponse} for the intent and get the user's name from the Session.
-         * 
-         * @param intent
-         *            intent for the request
-         * @return SpeechletResponse spoken and visual response for the intent
-         */
-        private SpeechletResponse GetNameFromSessionAndSayHello(Intent intent, Session session)
-        {
-            string speechOutput = "";
-            bool shouldEndSession = false;
-
-            // Get the user's name from the session.
-            string name = (String)session.Attributes[NAME_KEY];
-
-            // Check to make sure user's name is set in the session.
-            if (!String.IsNullOrEmpty(name))
-            {
-                speechOutput = String.Format("Your name is {0}, goodbye", name);
-                shouldEndSession = true;
-            }
-            else
-            {
-                // Since the user's name is not set render an error message.
-                speechOutput = "I'm not sure what your name is, you can say, my name is Sam";
-            }
-
-            return BuildSpeechletResponse(intent.Name, speechOutput, shouldEndSession);
-        }
-
-
-        /**
-         * Creates and returns the visual and spoken response with shouldEndSession flag
-         * 
-         * @param title
-         *            title for the companion application home card
-         * @param output
-         *            output content for speech and companion application home card
-         * @param shouldEndSession
-         *            should the session be closed
-         * @return SpeechletResponse spoken and visual response for the given input
-         */
         private SpeechletResponse BuildSpeechletResponse(string title, string output, bool shouldEndSession)
         {
-            // Create the Simple card content.
-            SimpleCard card = new SimpleCard();
-            card.Title = String.Format("SessionSpeechlet - {0}", title);
-            card.Subtitle = String.Format("SessionSpeechlet - Sub Title");
-            card.Content = String.Format("SessionSpeechlet - {0}", output);
+            // Create the Simple card content
+            var card = new SimpleCard
+            {
+                Title = $"SessionSpeechlet - {title}",
+                Content = $"SessionSpeechlet - {output}"
+            };
 
-            // Create the plain text output.
-            PlainTextOutputSpeech speech = new PlainTextOutputSpeech();
-            speech.Text = output;
+            // Create the plain text output
+            var speech = new PlainTextOutputSpeech { Text = output };
 
-            // Create the speechlet response.
-            SpeechletResponse response = new SpeechletResponse();
-            response.ShouldEndSession = shouldEndSession;
-            response.OutputSpeech = speech;
-            response.Card = card;
+            // Create the speechlet response
+            var response = new SpeechletResponse
+            {
+                ShouldEndSession = shouldEndSession,
+                OutputSpeech = speech,
+                Card = card
+            };
             return response;
+        }
+
+        public override Task OnSessionEndedAsync(SessionEndedRequest sessionEndedRequest, Session session)
+        {
+            throw new NotImplementedException();
+        }
+
+        public override Task OnSessionStartedAsync(SessionStartedRequest sessionStartedRequest, Session session)
+        {
+            throw new NotImplementedException();
         }
     }
 }
