@@ -9,7 +9,7 @@ namespace BusInfo
 
         public Task<string> GetJsonForTimeZoneFromLatLongAsync(string lat, string lon)
         {
-            var json = LoadJson(@"c:\users\kaseyu\documents\visual studio 2017\Projects\ClassLibrary1\UnitTestProject1\TimeZone.json");
+            var json = LoadJson(@"C:\Users\kaseyu\Source\Repos\TransitKing\UnitTestProject1\TimeZone.json");
             return Task.FromResult(json);
         }
     }
